Organized persistence classes under params/persistence dir
