+++ entropylab/pipeline/params/persistence/persistence.py
@@ -1,10 +1,9 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import ContextManager
 
 
-class Persistence(ABC, ContextManager):
+class Persistence(ABC):
     @abstractmethod
     def close(self):
         pass
