+++ entropylab/pipeline/tests/test_param_store.py
@@ -14,10 +14,10 @@ from entropylab.pipeline.api.in_process_param_store import (
     InProcessParamStore,
     MergeStrategy,
 )
-from entropylab.pipeline.api.migrations import fix_param_qualified_name, \
-    migrate_param_store_0_1_to_0_2
 from entropylab.pipeline.api.param_store import Param, LOCAL_TZ, _ns_to_datetime
-from entropylab.pipeline.api.tinydb_persistence import (
+from entropylab.pipeline.params.migrations import fix_param_qualified_name, \
+    migrate_param_store_0_1_to_0_2
+from entropylab.pipeline.params.persistence.tinydb_persistence import (
     Metadata,
     JSONPickleStorage,
     _set_version,
